requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
redis[hiredis]==5.0.1
websockets==12.0
httpx==0.26.0
python-multipart==0.0.9
pydantic-settings==2.1.0
pytest==7.4.4
pytest-asyncio==0.23.4
minio==7.2.4
Pillow
